pyproject.toml
[tool.poetry]
name = "rl-model"
version = "0.1.0"
description = ""
authors = ["litmus-zhang <abdulsalamlukmon9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mock-gpio = "^0.1.8"
ale-py = "^0.9.0"
gymnasium = {extras = ["box2d"], version = "^0.29.1"}
matplotlib = "^3.9.1"
seaborn = "^0.13.2"
tqdm = "^4.66.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

